refactor: 카드 유효성 검증 코드 utils 폴더로 이동
